requirements.txt
ollama==0.5.1              # Latest as of 2025
chromadb==1.0.15           # Latest as of July 2025  
textual==0.47.1            # Modern TUI framework
pydantic>=2.9.0            # Data validation (required by ollama)
pyyaml==6.0.1              # Config files
structlog==24.2.0          # Structured logging
python-dotenv==1.0.1       # Environment variables
aiosqlite==0.20.0          # Async SQLite for logging
prometheus-client==0.20.0  # Metrics tracking
pytest==8.2.2              # Testing
pytest-asyncio==0.23.7     # Async testing